@@ -26,15 +26,15 @@ repositories {
 }
 
 android {
-    compileSdkVersion 27
+    compileSdkVersion 28
     buildToolsVersion '28.0.2'
     defaultConfig {
         applicationId "korablique.recipecalculator"
         minSdkVersion 21
-        targetSdkVersion 27
+        targetSdkVersion 28
         versionCode 12
         versionName "18.02.04"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
     buildTypes {
         debug {
@@ -54,17 +54,17 @@ android {
 
 dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
-    androidTestCompile('com.android.support.test.espresso:espresso-core:3.0.1', {
+    androidTestCompile('androidx.test.espresso:espresso-core:3.1.1', {
         exclude group: 'com.android.support', module: 'support-annotations'
     })
-    androidTestCompile('com.android.support.test.espresso:espresso-contrib:3.0.1', {
+    androidTestCompile('androidx.test.espresso:espresso-contrib:3.1.1', {
         exclude group: 'com.android.support', module: 'support-annotations'
         exclude group: 'com.android.support', module: 'support-v4'
         exclude group: 'com.android.support', module: 'appcompat-v7'
         exclude group: 'com.android.support', module: 'design'
         exclude group: 'com.android.support', module: 'recyclerview-v7'
     })
-    androidTestImplementation 'com.android.support.test.espresso:espresso-intents:3.0.1'
+    androidTestImplementation 'androidx.test.espresso:espresso-intents:3.1.1'
 
     androidTestCompile('com.schibsted.spain:barista:2.2.0') {
         exclude group: 'com.android.support'
@@ -78,10 +78,10 @@ dependencies {
     compile('com.crashlytics.sdk.android:crashlytics:2.8.0@aar') {
         transitive = true
     }
-    compile 'com.android.support:appcompat-v7:27.1.0'
-    compile 'com.android.support:recyclerview-v7:27.1.0'
-    compile 'com.android.support:design:27.1.0'
-    compile 'com.android.support.constraint:constraint-layout:1.1.0'
+    compile 'androidx.appcompat:appcompat:1.1.0-alpha01'
+    compile 'androidx.recyclerview:recyclerview:1.1.0-alpha01'
+    compile 'com.google.android.material:material:1.1.0-alpha02'
+    compile 'androidx.constraintlayout:constraintlayout:2.0.0-alpha3'
     compile 'com.tapadoo.android:alerter:2.0.1'
     implementation project(':libs:floatingsearchview:library')
     implementation 'com.mikhaellopez:circularprogressbar:2.0.0'
@@ -99,5 +99,5 @@ dependencies {
     testCompile 'org.mockito:mockito-android:2.19.0'
     testCompile 'org.robolectric:robolectric:3.8'
     compile "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
-    testCompile "com.android.support.test:runner:1.0.2"
+    testCompile "androidx.test:runner:1.1.1"
 }
