@@ -1,9 +1,9 @@
 package korablique.recipecalculator.database;
 
 import android.content.Context;
-import android.support.test.InstrumentationRegistry;
-import android.support.test.filters.LargeTest;
-import android.support.test.runner.AndroidJUnit4;
+import androidx.test.InstrumentationRegistry;
+import androidx.test.filters.LargeTest;
+import androidx.test.runner.AndroidJUnit4;
 import android.util.MutableBoolean;
 
 import org.junit.Assert;
