@@ -1,7 +1,7 @@
 package korablique.recipecalculator.util;
 
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
 
 import java.lang.reflect.Field;
 import java.util.ArrayList;
