@@ -1,6 +1,6 @@
 package korablique.recipecalculator.base;
 
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 
 import java.util.NoSuchElementException;
 
