@@ -1,6 +1,6 @@
 package korablique.recipecalculator.dagger;
 
-import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
 
 import dagger.android.AndroidInjection;
 import dagger.android.support.AndroidSupportInjection;
