@@ -1,6 +1,6 @@
 package korablique.recipecalculator.dagger;
 
-import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
 
 import korablique.recipecalculator.BroccalcApplication;
 import korablique.recipecalculator.base.BaseActivity;
