@@ -2,7 +2,7 @@ package korablique.recipecalculator.model;
 
 import android.os.Parcel;
 import android.os.Parcelable;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import korablique.recipecalculator.FloatUtils;
 
