@@ -1,7 +1,7 @@
 package korablique.recipecalculator.ui.bucketlist;
 
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.helper.ItemTouchHelper;
+import androidx.recyclerview.widget.RecyclerView;
+import androidx.recyclerview.widget.ItemTouchHelper;
 
 
 public abstract class OnSwipeItemCallback extends ItemTouchHelper.SimpleCallback {
