@@ -2,7 +2,7 @@ package korablique.recipecalculator.ui.mainscreen;
 
 import android.content.Context;
 import android.content.Intent;
-import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
 
 import javax.inject.Inject;
 
