@@ -1,7 +1,7 @@
 package korablique.recipecalculator.ui.nestingadapters;
 
 
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.ViewGroup;
 
 import java.util.ArrayList;
