@@ -1,13 +1,12 @@
 package korablique.recipecalculator.ui.nestingadapters;
 
-import android.support.annotation.LayoutRes;
-import android.support.annotation.Nullable;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.LayoutRes;
+import androidx.annotation.Nullable;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-import korablique.recipecalculator.R;
 import korablique.recipecalculator.ui.MyViewHolder;
 
 public class SingleItemAdapterChild extends AdapterChild {
